Extract answer list into component
